# src/main/resources/static/swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "PeerFrame API",
		"version": "0.1.0",
		"description": "Local PeerFrame API."
	},
	"basePath": "/api",
	"paths": {
		"/providers": {
			"get": {
				"tags": [
					"Provider"
				],
				"responses": {
					"200": {
						"description": "Successful operation.",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Provider"
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Provider"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Provider"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation.",
						"schema": {
							"$ref": "#/definitions/Provider"
						}
					}
				}
			}
		},
		"/providers/{providerId}": {
			"put": {
				"tags": [
					"Provider"
				],
				"parameters": [
					{
						"in": "path",
						"name": "providerId",
						"type": "string",
						"required": true
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Provider"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation.",
						"schema": {
							"$ref": "#/definitions/Provider"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Provider"
				],
				"parameters": [
					{
						"in": "path",
						"name": "providerId",
						"type": "string",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation."
					}
				}
			}
		},
...
	},
	"definitions": {
		"Provider": {
			"description": "Generic provider resource.",
			"properties": {
				"id": {
					"type": "string",
					"description": "Unique Provider UUID."
				},
				"providerType": {
					"type": "string",
					"description": "Provider type (Smugmug, GooglePhotos, etc).",
					"enum": [
						"SmugmugProvider",
						"GooglePhotosProvider"
					]
				},
				"name": {
					"type": "string",
					"description": "Name of the provider resource."
				}
			},
			"discriminator": "providerType",
			"required": [
				"providerType",
				"name"
			]
		},
...
	}
}